+++ cli/tests/test_diffedit_command.rs
@@ -394,7 +394,7 @@ fn test_diffedit_merge() {
     A file3
     "###);
     assert!(!repo_path.join("file1").exists());
-    let stdout = test_env.jj_cmd_success(&repo_path, &["print", "file2"]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["file", "print", "file2"]);
     insta::assert_snapshot!(stdout, @r###"
     <<<<<<< Conflict 1 of 1
     %%%%%%% Changes from base to side #1